Add files via upload
This is a small Python project using Tkinter library, this is a Youtube watchlater organizer, where you can add a particular video url, which you have planned to watch later, you just have to copy the url of youtube video and press the Paste button on the screen and then press Add to add the link, refresh button simply displays the existing list of videos that are well organized with the channel name, the title of the video and the date added. i think the ui is great, considering the limitations of using tkinter.